move export headers to api dir
