+++ src/api/iptux-core/meson.build
@@ -1,0 +1,10 @@
+header_files = files([
+    'CoreThread.h',
+    'Event.h',
+    'Exception.h',
+    'IptuxConfig.h',
+    'Models.h',
+    'ProgramData.h',
+    'TransFileModel.h',
+])
+install_headers(header_files, subdir: 'iptux-core')
