+++ src/meson.build
@@ -38,6 +38,7 @@ configure_file(input : 'TestConfig.h.in',
                output : 'TestConfig.h',
                configuration : test_conf_data)
 
+subdir('api')
 subdir('iptux-utils')
 subdir('iptux-core')
 subdir('iptux')
